Create Invoke-SeasonEpisodeNaming.ps1 (#17)
Create a script to help with mass renaming files ripped via MakeMKV to
the format S00E00 by giving the Season and Start Episode and then
assuming that the order of the files (IE C_T01.mkv, C_T02.mkv) will be
the order of the Episodes. It will also rename the parent folder to the
format `Season.SeasonNumber` to assist Filebot.