Create Script to Query Jellyfin for Movies (#37)
This code was created using ChatGPT.

Instructed ChatGPT to create a PowerShell script that leveraged the
Jellyfin REST API to grab a list of movies with their titles in the
format of `Title (Year)` (if the year is present), along with the MPAA
Rating, and the Synopsis.

Out of the box this somewhat worked, however Jellyfin utilizes UTF-8
encoding which was not coming through. Reprompting ChatGPT made the
suggestion to change how Export worked, it does not seem the cleanest
but does work.